fix&create&delete models

+++ src/com/apmath/loans/domain/models/loans/Loan.kt
@@ -12,7 +12,7 @@ class Loan(
 
     override val amount: Money,
     override val term: Int,
-    override val interest: Int,
+    override val interest: Float,
     override val currency: Currency,
     override val date: String,
